Move job files to job package
