+++ molgenis-app/src/main/java/org/molgenis/app/WebAppPermissionRegistry.java
@@ -48,7 +48,7 @@ import org.molgenis.datarowedit.controller.DataRowEditController;
 import org.molgenis.metadata.manager.controller.MetadataManagerController;
 import org.molgenis.navigator.NavigatorController;
 import org.molgenis.navigator.copy.job.ResourceCopyJobExecutionMetadata;
-import org.molgenis.navigator.delete.ResourceDeleteJobExecutionMetadata;
+import org.molgenis.navigator.delete.job.ResourceDeleteJobExecutionMetadata;
 import org.molgenis.questionnaires.controller.QuestionnaireController;
 import org.molgenis.searchall.controller.SearchAllPluginController;
 import org.molgenis.security.core.PermissionSet;
