+++ molgenis-navigator/src/main/java/org/molgenis/navigator/NavigatorServiceImpl.java
@@ -24,8 +24,8 @@ import org.molgenis.jobs.JobExecutor;
 import org.molgenis.jobs.model.JobExecution;
 import org.molgenis.navigator.copy.job.ResourceCopyJobExecution;
 import org.molgenis.navigator.copy.job.ResourceCopyJobExecutionFactory;
-import org.molgenis.navigator.delete.ResourceDeleteJobExecution;
-import org.molgenis.navigator.delete.ResourceDeleteJobExecutionFactory;
+import org.molgenis.navigator.delete.job.ResourceDeleteJobExecution;
+import org.molgenis.navigator.delete.job.ResourceDeleteJobExecutionFactory;
 import org.molgenis.navigator.download.job.ResourceDownloadJobExecution;
 import org.molgenis.navigator.download.job.ResourceDownloadJobExecutionFactory;
 import org.molgenis.navigator.model.Resource;
