+++ molgenis-navigator/src/test/java/org/molgenis/navigator/NavigatorServiceImplTest.java
@@ -30,8 +30,8 @@ import org.molgenis.data.meta.model.PackageMetadata;
 import org.molgenis.jobs.JobExecutor;
 import org.molgenis.navigator.copy.job.ResourceCopyJobExecution;
 import org.molgenis.navigator.copy.job.ResourceCopyJobExecutionFactory;
-import org.molgenis.navigator.delete.ResourceDeleteJobExecution;
-import org.molgenis.navigator.delete.ResourceDeleteJobExecutionFactory;
+import org.molgenis.navigator.delete.job.ResourceDeleteJobExecution;
+import org.molgenis.navigator.delete.job.ResourceDeleteJobExecutionFactory;
 import org.molgenis.navigator.download.job.ResourceDownloadJobExecution;
 import org.molgenis.navigator.download.job.ResourceDownloadJobExecutionFactory;
 import org.molgenis.navigator.model.Resource;
